# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CASE-11",
        "description": "Bu API'nin genel dokümantasyonu.",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/blog_list": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get all blogs",
                "operationId": "81e1c9532026a6291ace2552aa8c0cfe",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter blogs by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of blogs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Blog"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No blogs found"
                    }
                }
            }
        },
        "/api/add_blog": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Add a new blog",
                "operationId": "69a47b0a6293dba8df4e2ddc9930fbd7",
                "requestBody": {
                    "description": "Blog data",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "Category ID",
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "header": {
                                        "description": "Blog header",
                                        "type": "string",
                                        "example": "güncelleme2"
                                    },
                                    "desc_short": {
                                        "description": "Short description",
                                        "type": "string",
                                        "example": "güncelleme2"
                                    },
                                    "desc_long": {
                                        "description": "Long description",
                                        "type": "string",
                                        "example": "güncelleme2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Blog added successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create a new user",
                "operationId": "a39949df013c1eb78ff8bd8f481d408a",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and generate token",
                "operationId": "c5e838284c7f343f0e5e5fd589161c96",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Blog": {
                "properties": {
                    "id": {
                        "description": "Blog ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "category_id": {
                        "description": "Category ID of the blog",
                        "type": "integer"
                    },
                    "header": {
                        "description": "Header of the blog",
                        "type": "string"
                    },
                    "desc_short": {
                        "description": "Short description of the blog",
                        "type": "string"
                    },
                    "desc_long": {
                        "description": "Long description of the blog",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Date and time when the blog was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date and time when the blog was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Date and time when the blog was soft deleted",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    }
}